Save the original model config in train script
The train script removes the "oxe_kwargs" key in the config, and add
"dataset_kwargs_list" & "sample_weights" to the config. Such behavior causes
a `TypeError: Object of type function is not JSON serializable(base)` error
when saving model checkpoint, because the `standardize_fn` we put into the
model config is not serializable. This commit fixes the issue by preserving the
unmodified config.